+++ util/opentelemetry-util-genai/src/opentelemetry/util/genai/client.py
