+++ util/opentelemetry-util-genai/tests/test_utils.py
@@ -8,7 +8,7 @@ from opentelemetry.sdk.trace.export import SimpleSpanProcessor
 from opentelemetry.sdk.trace.export.in_memory_span_exporter import (
     InMemorySpanExporter,
 )
-from opentelemetry.util.genai.api import (
+from opentelemetry.util.genai.client import (
     llm_start,
     llm_stop,
 )
